JavaFX-Maven-Apache
With this application, you can select a specific part of a movie subtitle and perform various actions. You have the option to search for a translation using Google Translate or check the definition using Oxford Dictionary. Additionally, you can explore how the selected sentence is pronounced in other movies on the Yarnn website. Furthermore, you have the ability to add specific words to an Excel file for future training purposes.